Cast FAT file size to off_t in stat
Make the `st_size` assignment explicit by casting `fno.fsize` to `off_t` in `stat()`. This aligns the file-size mapping with the target POSIX type and avoids implicit conversion issues.

+++ lib-fatfs/src/stat.cpp
@@ -78,7 +78,7 @@ int stat(const char* path, struct stat* buf) { // NOLINT
         return -1; // POSIX returns -1 on error
     }
 
-    buf->st_size = fno.fsize; // File size mapping
+    buf->st_size = static_cast<off_t>(fno.fsize);
 
     auto epoch = FatTimeToUnix(fno.fdate, fno.ftime);
     buf->st_mtime = epoch;
